Move the highlight flashcard endpoint to a flashcards-tagged router
POST /highlights/{highlight_id}/flashcards was defined inside the reading
highlights router, whose APIRouter carries tags=["highlights"]. Every other
flashcard endpoint is tagged "flashcards" — six routers across four different
prefixes (/books, /notes, /chapters, /flashcards) all agree on that tag — so
this one endpoint was the sole outlier.

The tag is not cosmetic. The frontend generates its client with orval in
tags-split mode, which groups modules by OpenAPI tag rather than URL prefix, so
this endpoint was the only reason a caller creating a flashcard had to import
from the generated highlights module. It now lands with the rest of them.

The new router mirrors note_flashcards.py, which has the same shape for notes.
Its prefix plus route path reproduce the previous absolute path exactly, and the
handler name is unchanged, so FastAPI derives an identical operationId — the
generated hook keeps its name and only changes file. The OpenAPI document
differs in one field: tags.
